fix(viewer): invalidate binary viewer byte cache on fs_changed
External modifications to images/PDFs previously left the viewer byte
cache stale forever. Wire the cache into the fs_changed event: add
invalidateFsChanged/fsChangeAffectsPath to viewerByteCache, invalidate
globally from App, and auto-reload mounted ImageViewer/PdfViewer via a
reloadKey listener. Add 13 tests (9 unit + 4 component).

Also clean up 16 non-git-module clippy warnings in src-tauri
(mechanical, no behavior change) and add .github/workflows/ci.yml
(frontend tsc/vitest + rust clippy/test on push and pull request).

+++ src-tauri/src/fs/search.rs
@@ -574,7 +574,7 @@ pub(crate) fn compute_file_replacement(
     // char offsets). If the target doesn't resolve to a real match, the
     // file produces no replacement.
     let target_match_line = target.and_then(|t| if t.line >= 1 { Some(t.line as usize) } else { None });
-    let target_match_col = target.and_then(|t| if t.column >= 1 { Some(t.column as u32) } else { None });
+    let target_match_col = target.and_then(|t| if t.column >= 1 { Some(t.column) } else { None });
 
     let mut rebuilt = String::with_capacity(text.len());
     let mut total_matches = 0usize;
